refactor(wallet): split sso.tsx into co-located subcomponents
Extract the seven inline SSO subcomponents (MerchantIcon, SsoSubtitle,
SsoDisclaimer, SsoActions, ContinueAsSession + PhonePairingAction /
BackToSessionAction) into siblings under component/Sso/, the useLoginDemo
mutation into hook/, and the pure shortenAddress helper into utils/ with a
unit test. sso.tsx keeps only the Route definition and the Sso orchestrator.
Behavior-preserving; no rendered-output or public-API change.

FRA-264
